+++ package.json
@@ -4,7 +4,7 @@
   "description": "TDD",
   "main": "app.js",
   "scripts": {
-    "test": "eslint *.js && mocha -R nyan *.test.js",
+    "test": "eslint *.js && mocha -R nyan tests",
     "watch": "npm test -- --watch"
   },
   "repository": {
